Extract sandbox backends into libs/vs-sandbox workspace package
