+++ libs/vs-sandbox/tests/test_docker_sandbox.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 
 import pytest
 
-from vibesys.sandbox.docker_sandbox import DockerSandbox
+from vs_sandbox.docker_sandbox import DockerSandbox
 
 
 @pytest.fixture()
@@ -467,7 +467,7 @@ class TestCleanupOnExit:
     @pytest.fixture(autouse=True)
     def _clear_live_containers(self):
         """Isolate the global _live_containers registry between tests."""
-        from vibesys.sandbox.docker_sandbox import _live_containers
+        from vs_sandbox.docker_sandbox import _live_containers
 
         _live_containers.clear()
         yield
@@ -475,7 +475,7 @@ class TestCleanupOnExit:
 
     @patch("subprocess.run")
     def test_live_containers_tracked(self, mock_run, sandbox):
-        from vibesys.sandbox.docker_sandbox import _live_containers
+        from vs_sandbox.docker_sandbox import _live_containers
 
         mock_run.return_value = subprocess.CompletedProcess(
             args=[], returncode=0, stdout="abc123\n", stderr=""
@@ -489,7 +489,7 @@ class TestCleanupOnExit:
 
     @patch("subprocess.run")
     def test_cleanup_containers_stops_all(self, mock_run, sandbox):
-        from vibesys.sandbox.docker_sandbox import _cleanup_containers, _live_containers
+        from vs_sandbox.docker_sandbox import _cleanup_containers, _live_containers
 
         mock_run.return_value = subprocess.CompletedProcess(
             args=[], returncode=0, stdout="container_xyz\n", stderr=""
