+++ libs/vs-sandbox/tests/test_modal_model_setup.py
@@ -7,7 +7,7 @@ import pytest
 
 class TestVolumeNameFor:
     def test_sanitizes_slash_and_case(self):
-        from vibesys.sandbox.modal_model_setup import _volume_name_for
+        from vs_sandbox.modal_model_setup import _volume_name_for
 
         assert (
             _volume_name_for("meta-llama/Llama-3.1-8B-Instruct")
@@ -15,7 +15,7 @@ class TestVolumeNameFor:
         )
 
     def test_handles_colons_and_underscores(self):
-        from vibesys.sandbox.modal_model_setup import _volume_name_for
+        from vs_sandbox.modal_model_setup import _volume_name_for
 
         assert (
             _volume_name_for("openai/whisper_large-v3") == "vibesys-model-openai-whisper-large-v3"
@@ -55,7 +55,7 @@ def mock_modal(monkeypatch):
 
 class TestEnsureModelVolume:
     def test_skips_upload_when_sentinel_present(self, mock_modal):
-        from vibesys.sandbox.modal_model_setup import (
+        from vs_sandbox.modal_model_setup import (
             _READY_SENTINEL,
             ensure_model_volume,
         )
@@ -74,7 +74,7 @@ class TestEnsureModelVolume:
         assert any("ready" in line for line in logs)
 
     def test_triggers_upload_when_volume_empty(self, mock_modal):
-        from vibesys.sandbox.modal_model_setup import ensure_model_volume
+        from vs_sandbox.modal_model_setup import ensure_model_volume
 
         mock_modal["volume"].listdir.return_value = []
 
@@ -86,7 +86,7 @@ class TestEnsureModelVolume:
         assert any("populating" in line for line in logs)
 
     def test_triggers_upload_when_listdir_raises(self, mock_modal):
-        from vibesys.sandbox.modal_model_setup import ensure_model_volume
+        from vs_sandbox.modal_model_setup import ensure_model_volume
 
         mock_modal["volume"].listdir.side_effect = RuntimeError("not found")
 
@@ -98,7 +98,7 @@ class TestEnsureModelVolume:
     def test_forwards_hf_token_as_secret(self, mock_modal):
         import modal
 
-        from vibesys.sandbox.modal_model_setup import ensure_model_volume
+        from vs_sandbox.modal_model_setup import ensure_model_volume
 
         mock_modal["volume"].listdir.return_value = []
 
@@ -109,7 +109,7 @@ class TestEnsureModelVolume:
     def test_reads_hf_token_from_env(self, mock_modal, monkeypatch):
         import modal
 
-        from vibesys.sandbox.modal_model_setup import ensure_model_volume
+        from vs_sandbox.modal_model_setup import ensure_model_volume
 
         mock_modal["volume"].listdir.return_value = []
         monkeypatch.setenv("HF_TOKEN", "from_env")
