+++ libs/vs-sandbox/tests/test_modal_sandbox.py
@@ -10,7 +10,7 @@ def mock_modal(monkeypatch):
     """Patch modal.Sandbox.create, Image.from_registry, App.lookup, Volume.from_name/delete."""
     import modal
 
-    from vibesys.sandbox.modal_sandbox import _live_sandboxes
+    from vs_sandbox.modal_sandbox import _live_sandboxes
 
     _live_sandboxes.clear()
 
@@ -54,7 +54,7 @@ def mock_modal(monkeypatch):
 
 @pytest.fixture()
 def sandbox(tmp_path, mock_modal):
-    from vibesys.sandbox.modal_sandbox import ModalSandbox
+    from vs_sandbox.modal_sandbox import ModalSandbox
 
     ws = tmp_path / "workspace"
     ws.mkdir()
@@ -77,7 +77,7 @@ class TestVpath:
         assert sandbox._vpath("/workspace/foo") == "/workspace/foo"
 
     def test_preserves_passthrough(self, tmp_path, mock_modal):
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         sb = ModalSandbox(
             host_workspace=str(tmp_path),
@@ -105,7 +105,7 @@ class TestStart:
     def test_start_mounts_model_volume_when_name_given(self, tmp_path, mock_modal):
         import modal
 
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         with ModalSandbox(
             host_workspace=str(tmp_path),
@@ -120,7 +120,7 @@ class TestStart:
         tmp_path,
         mock_modal,
     ):
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         ws = tmp_path / "ws"
         ws.mkdir()
@@ -141,7 +141,7 @@ class TestStart:
         tmp_path,
         mock_modal,
     ):
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         ws = tmp_path / "ws"
         ws.mkdir()
@@ -159,7 +159,7 @@ class TestStart:
         tmp_path,
         mock_modal,
     ):
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         ws = tmp_path / "ws"
         ws.mkdir()
@@ -182,7 +182,7 @@ class TestStart:
         tmp_path,
         mock_modal,
     ):
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         ws = tmp_path / "ws"
         ws.mkdir()
@@ -212,7 +212,7 @@ class TestStart:
         tmp_path,
         mock_modal,
     ):
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         ws = tmp_path / "ws"
         ws.mkdir()
@@ -299,7 +299,7 @@ class TestTransientRetry:
     def test_is_transient_detects_dns_and_connection_errors(self):
         import socket
 
-        from vibesys.sandbox.modal_sandbox import _is_transient
+        from vs_sandbox.modal_sandbox import _is_transient
 
         assert _is_transient(socket.gaierror(-2, "Name or service not known"))
         assert _is_transient(ConnectionResetError("Connection reset"))
@@ -316,10 +316,10 @@ class TestTransientRetry:
     def test_retry_recovers_after_transient_error(self, monkeypatch):
         import socket
 
-        from vibesys.sandbox.modal_sandbox import _retry_transient
+        from vs_sandbox.modal_sandbox import _retry_transient
 
         # Avoid actual sleeping in the test.
-        monkeypatch.setattr("vibesys.sandbox.modal_sandbox.time.sleep", lambda _: None)
+        monkeypatch.setattr("vs_sandbox.modal_sandbox.time.sleep", lambda _: None)
 
         calls = {"n": 0}
 
@@ -333,7 +333,7 @@ class TestTransientRetry:
         assert calls["n"] == 3
 
     def test_retry_reraises_non_transient_immediately(self):
-        from vibesys.sandbox.modal_sandbox import _retry_transient
+        from vs_sandbox.modal_sandbox import _retry_transient
 
         calls = {"n": 0}
 
@@ -348,9 +348,9 @@ class TestTransientRetry:
     def test_retry_gives_up_after_max_attempts(self, monkeypatch):
         import socket
 
-        from vibesys.sandbox.modal_sandbox import _retry_transient
+        from vs_sandbox.modal_sandbox import _retry_transient
 
-        monkeypatch.setattr("vibesys.sandbox.modal_sandbox.time.sleep", lambda _: None)
+        monkeypatch.setattr("vs_sandbox.modal_sandbox.time.sleep", lambda _: None)
 
         calls = {"n": 0}
 
@@ -366,7 +366,7 @@ class TestTransientRetry:
         """Integration: a flaky sandbox.exec should be transparently retried."""
         import socket
 
-        monkeypatch.setattr("vibesys.sandbox.modal_sandbox.time.sleep", lambda _: None)
+        monkeypatch.setattr("vs_sandbox.modal_sandbox.time.sleep", lambda _: None)
         sandbox.start()
 
         proc_ok = MagicMock()
@@ -497,7 +497,7 @@ class TestSandboxFallbackRestart:
     """Verify the sandbox-dead fallback path: recreate + retry, preserving the volume."""
 
     def test_is_sandbox_dead_recognizes_shutdown_messages(self):
-        from vibesys.sandbox.modal_sandbox import _is_sandbox_dead
+        from vs_sandbox.modal_sandbox import _is_sandbox_dead
 
         assert _is_sandbox_dead(RuntimeError("Sandbox has already shut down"))
         assert _is_sandbox_dead(Exception("Sandbox has exited"))
@@ -515,7 +515,7 @@ class TestSandboxFallbackRestart:
         monkeypatch,
     ):
         """A dead sandbox should be recreated once, then the command re-run."""
-        monkeypatch.setattr("vibesys.sandbox.modal_sandbox.time.sleep", lambda _: None)
+        monkeypatch.setattr("vs_sandbox.modal_sandbox.time.sleep", lambda _: None)
         sandbox.start()
 
         proc_ok = MagicMock()
@@ -541,7 +541,7 @@ class TestSandboxFallbackRestart:
 
     def test_restart_attempts_capped_by_max(self, sandbox, mock_modal, monkeypatch):
         """After max_restart_attempts, further restarts are refused."""
-        monkeypatch.setattr("vibesys.sandbox.modal_sandbox.time.sleep", lambda _: None)
+        monkeypatch.setattr("vs_sandbox.modal_sandbox.time.sleep", lambda _: None)
         sandbox._max_restart_attempts = 1
         sandbox.start()
 
@@ -551,14 +551,14 @@ class TestSandboxFallbackRestart:
 
     def test_restart_can_be_disabled(self, tmp_path, mock_modal, monkeypatch):
         """enable_fallback_restart=False disables the recovery path entirely."""
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         with ModalSandbox(
             host_workspace=str(tmp_path),
             image="nvcr.io/nvidia/pytorch:25.04-py3",
             enable_fallback_restart=False,
         ) as sb:
-            monkeypatch.setattr("vibesys.sandbox.modal_sandbox.time.sleep", lambda _: None)
+            monkeypatch.setattr("vs_sandbox.modal_sandbox.time.sleep", lambda _: None)
 
             call = {"n": 0}
 
@@ -576,7 +576,7 @@ class TestSandboxFallbackRestart:
         """Auxiliary volumes like /draft_model should be added to volumes dict."""
         import modal
 
-        from vibesys.sandbox.modal_sandbox import ModalSandbox
+        from vs_sandbox.modal_sandbox import ModalSandbox
 
         with ModalSandbox(
             host_workspace=str(tmp_path),
