+++ tests/backends/test_gpu_monitor.py
@@ -270,7 +270,7 @@ class TestReselectGpu:
         """Build a minimal _RunContext-like object for reselect_gpu testing."""
         from vibesys.backends.cuda import CudaBackend
         from vibesys.context import _RunContext
-        from vibesys.sandbox.docker_sandbox import DockerSandbox
+        from vs_sandbox.docker_sandbox import DockerSandbox
 
         ctx = object.__new__(_RunContext)
         ctx.selected_gpu = selected_gpu
